+++ src/KeyVault/KeyVault/Commands/FullBackupRestore/RestoreAzureManagedHsm.cs
@@ -14,6 +14,9 @@ namespace Microsoft.Azure.Commands.KeyVault.Commands
         [Parameter(Mandatory = true, HelpMessage = "Folder name of the backup, e.g. 'mhsm-*-2020101309020403'.\nIt can also be nested such as 'backups/mhsm-*-2020101309020403'.")]
         public string BackupFolder { get; set; }
 
+        [Parameter(Mandatory = false, HelpMessage = "Key name to restore.")]
+        public string KeyName { get; set; }
+
         [Parameter(Mandatory = false, HelpMessage = "Return true when the HSM is restored.")]
         public SwitchParameter PassThru { get; set; }
 
@@ -23,14 +26,24 @@ namespace Microsoft.Azure.Commands.KeyVault.Commands
                     string.Format(Resources.DoFullRestore, StorageContainerUri),
                     HsmName, () =>
                     {
+                        var errorMsg = string.Format(Resources.FullRestoreFailed, HsmName);
                         try
                         {
-                            Track2DataClient.RestoreHsm(HsmName, StorageContainerUri, SasToken.ConvertToString(), BackupFolder);
+                            if(KeyName == null)
+                            {
+                                Track2DataClient.RestoreHsm(HsmName, StorageContainerUri, SasToken.ConvertToString(), BackupFolder);
+                            }
+                            else
+                            {
+                                Track2DataClient.SelectiveRestoreHsm(HsmName, KeyName,StorageContainerUri, SasToken.ConvertToString(), BackupFolder);
+                                errorMsg = string.Format(Resources.SelectiveRestoreFailed, KeyName, HsmName);
+                            }
                         }
                         catch (Exception ex)
                         {
-                            throw new Exception(string.Format(Resources.FullRestoreFailed, HsmName), ex);
+                            throw new Exception(errorMsg, ex);                            
                         }
+                        
                         if (PassThru)
                         {
                             WriteObject(true);